+++ packages/adapter-sql-js/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @powersync/adapter-sql-js
 
+## 0.0.21
+
+### Patch Changes
+
+- Updated dependencies [c787aaf]
+  - @powersync/shared-internals@1.1.1
+
+## 0.0.20
+
+### Patch Changes
+
+- 3d1fd84: Update PowerSync SQLite core extension to version 0.5.2.
+- Updated dependencies [3d1fd84]
+- Updated dependencies [669ba97]
+- Updated dependencies [bc50ecc]
+  - @powersync/shared-internals@1.1.0
+
 ## 0.0.19
 
 ### Patch Changes