+++ packages/nuxt/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @powersync/nuxt
 
+## 0.0.9
+
+### Patch Changes
+
+- 3d1fd84: Make `@journeyapps/wa-sqlite` a regular dependency instead of a dev-dependency.
+  When upgrading, you can remove that package from your dependencies if you don't import it directly in your project.
+- 3d1fd84: Update PowerSync SQLite core extension to version 0.5.2.
+- Updated dependencies [3d1fd84]
+- Updated dependencies [7f886e4]
+- Updated dependencies [3d1fd84]
+- Updated dependencies [669ba97]
+- Updated dependencies [8d18368]
+- Updated dependencies [8d18368]
+- Updated dependencies [bc50ecc]
+- Updated dependencies [d1b3622]
+  - @powersync/web@2.1.0
+  - @powersync/kysely-driver@2.0.1
+  - @powersync/shared-internals@1.1.0
+  - @powersync/vue@0.6.0
+
 ## 0.0.8
 
 ### Minor Changes